a bit of cleanup and correct int -> size_t for alloc for fasta

+++ RIsearch2/src/fasta.c
@@ -99,8 +99,8 @@ int fasta_read(fasta_t *ffp, char **ret_seq, char **ret_name, saidx64_t *ret_L)
 	char *s;
 	char *name;
 	char *seq;
-	int   n;
-	int   nalloc;
+	size_t   n;
+	size_t   nalloc;
 //	int   nonalpha = 0, modified = 0;
 	const char *valid = "AaCcGgUuNn";
 
@@ -153,7 +153,7 @@ int fasta_read(fasta_t *ffp, char **ret_seq, char **ret_name, saidx64_t *ret_L)
 				seq = realloc(seq, sizeof(char) * nalloc);
 				if (seq == NULL) {
 					fprintf(stderr, "%s\n", ffp->buffer);
-					fprintf(stderr, "%d - %d\n", n, nalloc);
+					fprintf(stderr, "%ld - %ld\n", n, nalloc);
 					fprintf(stderr, "Failed re-allocation in fasta read.\n");
 					abort();
 				}
